test save-file

+++ test/domain/use-case/save-file.use-case.test.ts
@@ -1,5 +1,5 @@
 import { SaveFile } from '../../../src/domain/use-cases/save-filte.use-case'
-import fs, { ftruncate }  from 'fs'
+import fs  from 'fs'
 
 describe('SaveFileUseCase', () => {
 
@@ -9,7 +9,11 @@ describe('SaveFileUseCase', () => {
 
     afterEach(() => {
         // clean up
-        fs.rmSync('outputs', { recursive: true })
+
+        const outputFolderExist = fs.existsSync('outputs')
+        
+        if(outputFolderExist) fs.rmSync('outputs', { recursive: true })
+        
     })
 
     it('should save file with default values', () => {
@@ -24,6 +28,33 @@ describe('SaveFileUseCase', () => {
         const result = saveFile.execute(options)
         const fileExist = fs.existsSync(filePath)
         const fileContent = fs.readFileSync(filePath, { encoding: 'utf-8' })
+
+        expect(result).toBeTruthy()
+        expect(fileExist).toBeTruthy()
+        expect(fileContent).toBe(options.fileContent)
+
+
+
+    })
+
+    it('should save file with custom values ', () => {
+
+        const options = {
+            fileContent: 'custom content',
+            fileDestination: 'custom-outputs',
+            fileName: 'custom-table-name'
+        }
+
+        const saveFile = new SaveFile()
+        const path = `outputs/${options.fileDestination}/${options.fileName}.txt`
+        const result = saveFile.execute({
+            ...options,
+            fileDestination: `outputs/${options.fileDestination}`
+        })
+
+
+        const fileExist = fs.existsSync(path)
+        const fileContent = fs.readFileSync(path, { encoding: 'utf-8' })
 
         expect(result).toBeTruthy()
         expect(fileExist).toBeTruthy()
